+++ packages/map-template/src/components/ChatButton/ChatButton.jsx
@@ -1,4 +1,5 @@
 import { createPortal } from 'react-dom';
+import i18n from 'i18next';
 import { useRecoilValue } from 'recoil';
 import PropTypes from 'prop-types';
 import chatHistoryState from '../../atoms/chatHistoryState';
@@ -46,8 +47,8 @@ function ChatButton({ pushAppView, currentAppView, appViews }) {
         <button 
             className="chat-button"
             onClick={handleOpenChat}
-            title="Open chat"
-            aria-label="Open chat"
+            title={i18n.t('Open chat')}
+            aria-label={i18n.t('Open chat')}
             type="button"
             style={{ '--chat-button-primary-color': primaryColor }}
         >